Update build type in imdb CMakeLists
The CMakeLists.txt file for the imdb component has been modified to set the CMAKE_BUILD_TYPE to "Release" from "Debug". This change will enable the component to be built in the release mode. It will help generating highly optimized executable and also potentially reduce runtime execution errors and crashes.

Signed-off-by: yiqiao <yimingqiao3163@gmail.com>